requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
paho-mqtt
python-dotenv